fix: delve
shoulde have fixed whatever is going on with delve, it should work now.
I've also added in another plugin, but it's commented out at the moment.
want to look into it a it more and make sure I'm setting it up
correctly, or that I even need it at all.