feat: implemented support for zig
I added zig and zls because I'm jumping on the hype train. Also, I added
a fix for the multiple offset_encoding warning that pops up whenever I
do stuff with C++. I don't remember if this issue is present in rust or zig.
Will check later if the issues are present and fix as necessary.