feat: diagnostics improvements
Some QOL plugins have been added focusing on diagnostic elements.
virtual text created by the lsp is now managed by a plugin, so the error
messages will no longer extend outside the view of the buffer. Also,
many plugins that have to do with diagnostics have been moved to a
single diagnostics lua file. Diagnostics of a file can now be viewed in
the oil buffer.
Not sure which plugin is causing it, will need to test; if there are
errors in a file, the diagnostics are viewed in the oil buffer,
and then gone back into the file to fix, the current line number is
highlighted, even in the oil buffer, and does not go away until
ending the neovim session.